settings now comes when you save settings

+++ src/main.py
@@ -2,6 +2,7 @@ import json
 import utils as ut
 import streamlit as st
 import pyperclip
+import os
 from time import perf_counter
 
 
@@ -31,16 +32,20 @@ def settings_management():
 
         with st.container(horizontal=True, horizontal_alignment='left'):
             if st.button("Save Settings"):
-                with open("../settings.json", "w") as f:
+                with open("settings.json", "w") as f:
                     json.dump(st.session_state.settings, f, indent=2)
+                st.toast("Settings Saved")
 
             if st.button("Set to Default Settings"):
-                with open("../default_settings.json", "r") as f:
+                with open("default_settings.json", "r") as f:
                     default_settings = json.load(f)
 
-                with open("../settings.json", "w") as f:
-                    json.dump(default_settings, f, indent=2)
-                    st.session_state.settings = default_settings
+                st.session_state.settings = default_settings
+
+                if os.path.exists("settings.json"):
+                    with open("settings.json", "w") as f:
+                        json.dump(default_settings, f, indent=2)
+
                 st.rerun()
 
 def display_transcribing(uploaded_file):
@@ -86,8 +91,12 @@ def display_result(result, time_taken):
 def main():
 
     if "settings" not in st.session_state:
-        with open("../settings.json", "r") as file:
-            st.session_state["settings"] = json.load(file)
+        if os.path.exists("settings.json"):
+            with open("settings.json", "r") as file:
+                st.session_state["settings"] = json.load(file)
+        else:
+            with open("default_settings.json", "r") as file:
+                st.session_state["settings"] = json.load(file)
 
     st.set_page_config("Transcribot", page_icon="🔉")
